requirements.txt
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.0
fonttools==4.61.1
fsspec==2025.12.0
imbalanced-learn==0.14.1
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.3.5
packaging==26.0
pandas==3.0.1
pillow==12.0.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
setuptools==70.2.0
six==1.17.0
sklearn-compat==0.1.5
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
